move acceptreject to it's own file
